pyproject.toml
[project]
name = "jobpilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datetime>=6.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "google-generativeai>=0.8.6",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
]
